+++ core/src/main/scala/restui/models/Service.scala
@@ -1,0 +1,4 @@
+package restui.models
+
+final case class Service(serviceName: String, file: OpenApiFile, metadata: Map[String, String] = Map.empty)
+final case class OpenApiFile(contentType: ContentType, content: String)
