+++ core/src/main/scala/restui/models/ServiceEvent.scala
@@ -1,0 +1,8 @@
+package restui.models
+
+sealed trait ServiceEvent
+
+object ServiceEvent{
+  final case class ServiceUp(service: Service)      extends ServiceEvent
+  final case class ServiceDown(serviceName: String) extends ServiceEvent
+}
