+++ rest-ui/src/main/scala/restui/server/service/EventSource.scala
@@ -1,13 +1,13 @@
 package restui.server.service
 
-import scala.concurrent.duration._
-
 import akka.NotUsed
 import akka.http.scaladsl.model.sse.ServerSentEvent
 import akka.stream.scaladsl.{BroadcastHub, Keep, RunnableGraph, Source, SourceQueueWithComplete}
 import akka.stream.{DelayOverflowStrategy, OverflowStrategy}
 import io.circe.syntax._
-import restui.server.http.Models._
+import restui.models.Event
+
+import scala.concurrent.duration._
 
 object EventSource {
   def createEventSource: RunnableGraph[(SourceQueueWithComplete[Event], Source[ServerSentEvent, NotUsed])] =
