+++ rest-ui/src/main/scala/restui/server/service/ServiceActor.scala
@@ -2,21 +2,20 @@ package restui.server.service
 
 import akka.actor.{Actor, ActorLogging, Props}
 import akka.stream.scaladsl.SourceQueueWithComplete
-import restui.providers.models._
-import restui.server.http.Models
+import restui.models._
 
-class ServiceActor(queue: SourceQueueWithComplete[Models.Event]) extends Actor with ActorLogging {
+class ServiceActor(queue: SourceQueueWithComplete[Event]) extends Actor with ActorLogging {
   import ServiceActor._
   override def receive: Receive = handleReceive(Map.empty)
 
   private def handleReceive(services: Map[String, Service]): Receive = {
-    case (provider: String, ServiceUp(service)) =>
+    case (provider: String, ServiceEvent.ServiceUp(service)) =>
       log.debug("{} got a new service", provider)
-      queue.offer(Models.ServiceUp(service.serviceName))
+      queue.offer(Event.ServiceUp(service.serviceName))
       context.become(handleReceive(services + (service.serviceName -> service)))
 
-    case (provider: String, ServiceDown(serviceName)) =>
-      queue.offer(Models.ServiceDown(serviceName))
+    case (provider: String, Event.ServiceDown(serviceName)) =>
+      queue.offer(Event.ServiceDown(serviceName))
       log.debug("{} removed a service", provider)
       context.become(handleReceive(services - serviceName))
 
@@ -27,7 +26,7 @@ class ServiceActor(queue: SourceQueueWithComplete[Models.Event]) extends Actor w
 }
 
 object ServiceActor {
-  def props(queue: SourceQueueWithComplete[Models.Event]): Props = Props(classOf[ServiceActor], queue)
+  def props(queue: SourceQueueWithComplete[Event]): Props = Props(classOf[ServiceActor], queue)
   case class Get(serviceName: String)
   case object GetAll
 }
